pyproject.toml
[tool.poetry]
name = "test-fastapi"
version = "0.1.0"
description = ""
authors = ["Michael Toohig <michael.toohig@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.95.0"
uvicorn = "^0.21.1"
SQLAlchemy = {version = "^2.0.7", extras = ["asyncio"]}
pydantic = "1.10.7"
aiosqlite = "^0.18.0"
Jinja2 = "^3.1.2"
loguru = "^0.6.0"
asgiref = "^3.6.0"
boussole = "^2.1.0"
httpx = {version = "^0.23.3", extras = ["http2"]}
beautifulsoup4 = "^4.12.0"
alembic = "^1.10.2"
Cerberus = "^1.3.4"
cerberus-list-schema = "^1.0.3"
python-dateutil = "^2.8.2"

[tool.poetry.dev-dependencies]

[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
pytest = "^7.3.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
